feat: implement ClientPage component and refactor Home component for improved structure and async data fetching
